Refactor alpaca.Utilities to util.TimeUtil
